requirements.txt
black
isort
flake8
python-dotenv
debugpy
pytest
tree-math
optax
matplotlib
numpy
jupyter
gitpython
opencv-python
scikit-learn
wandb
tqdm
hydra-core
hydra_colorlog
omegaconf